So I did the GUI stuff for a create_course, create_misc, course_view, misc_view, and progress_view.
Not sure what to do with the course_listview or misc_listview.

Also left out the date/time aspects of the course_view, misc_view for now.

Updated strings.xml but it was giving me some trouble.

Also added a method to the Course class to give the grade back as a string for ease of use.

-Brian